+++ lib/src/exception/internal_server_error.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'base_http_exception.dart';
 
-class InternalServerError extends BaseHttpException {
+class InternalServerError extends BaseHttpResponseException {
   InternalServerError(
       {required super.message,
       super.code = HttpStatus.internalServerError,