+++ lib/src/exception/unauthenticated.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'base_http_exception.dart';
 
-class Unauthenticated extends BaseHttpException {
+class Unauthenticated extends BaseHttpResponseException {
   Unauthenticated(
       {required super.message,
       super.code = HttpStatus.unauthorized,