+++ lib/vania.dart
@@ -56,3 +56,5 @@ export 'package:mailer/src/entities/attachment.dart';
 export 'src/utils/helper.dart';
 export 'src/env_handler/env.dart';
 export 'src/logger/logger.dart';
+
+export 'src/database/seeder/seeder.dart';